Implement read/write operations to the notmuch database
Keep only a single database connection open, close it when the app exits.

Split NotMuch::Message and NotMuch::Thread into their own files, because they
now contain additional information to manipulate those objects.

Most access to the notmuch database is done in notmuch.cc, thread.cc
or message.cc. The idea is to remove this logic from the views. This is not
done entirely, the search thread in search_view.cc still contains a fair bit
of logic which doesn't belong there.